Outsource SVG container into separate component
